Abstract base64 to buffer conversion logic
